requirements.txt
asgiref==3.8.1
Django==5.1
django-dump-load-utf8==0.0.4
django-filter==24.3
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
